refactor(desktop): extract text preview helpers
